+++ docs/Feedback.md
@@ -22,5 +22,5 @@
   - 해당 과정에서 문제점이 equals 메서드인지 의심이 되어 수정해보았음.
 
 ## Level 3 - JsonView 구현
-- [ ] 리팩터링 이후 실패하는 테스트 코드 개선
+- [x] 리팩터링 이후 실패하는 테스트 코드 개선
 - [x] JsonView 힌트 적용
