+++ mvc/src/test/java/nextstep/mvc/controller/AnnotationHandlerAdapterTest.java
@@ -1,10 +1,10 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import nextstep.mvc.controller.AnnotationHandlerAdapter;
 import nextstep.mvc.controller.HandlerExecution;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.web.annotation.Controller;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
